fix(streaming): isolate volatile ISR counter, atomic clear, 64-bit (#265)
Address Qodo review feedback on PR #266:

1. Isolate timerISRCalls as a separate `static volatile uint64_t
   gTimerISRCalls` global, distinct from gStreamStats. The previous
   in-struct field relied implicitly on the compiler not optimizing
   ISR-side accesses; making it volatile is the correct way to
   guarantee fresh reads from the snapshot side. (Importance 8)

2. Promote to uint64_t — at the ~90 kHz hardware ceiling, uint32 wraps
   in ~13 hours. uint64 takes ~6 million years and matches the existing
   totalSamplesStreamed/totalBytesStreamed fields. The non-atomic 64-bit
   read is safe because the snapshot is inside taskENTER_CRITICAL which
   blocks the timer ISR (priority 1).

3. Add configASSERT(uxInterruptNesting == 0) in Streaming_ClearStats so
   future ISR-context callers fail loudly instead of silently breaking
   on taskENTER_CRITICAL. (Importance 6)

4. Move Logger_ResetSessionOneShots() inside the critical section so
   the entire session-observability state is reset atomically — closes
   a window where an observer could see fresh stats but stale log
   one-shots. (Importance 4)

Hardware test: invariant TimerISRCalls == TotalSamples + QueueDropped
verified across 5 rates (1k, 11k, 30k, 100k, 200k Hz) on NQ1 USB.

Entire-Checkpoint: acaf7329ca1a

+++ firmware/src/services/streaming.h
@@ -116,7 +116,17 @@ typedef struct {
     // The invariant `timerISRCalls == totalSamplesStreamed + queueDroppedSamples`
     // should always hold during a session — every timer event becomes either
     // a successfully queued sample or a pool-exhaustion drop.
-    uint32_t timerISRCalls;          // Actual timer ISR entry count this session
+    //
+    // 64-bit so it never wraps in practice — at the ~90 kHz hardware ceiling
+    // it would take ~6 million years to overflow. Matches the other 64-bit
+    // session counters (totalSamplesStreamed, totalBytesStreamed).
+    //
+    // Storage note: this field is populated by Streaming_GetStats() from a
+    // separate `static volatile uint64_t gTimerISRCalls` global. The volatile
+    // global is the actual ISR-modified storage; the StreamingStats field is
+    // a snapshot copy taken inside taskENTER_CRITICAL (which makes the
+    // non-atomic 64-bit read coherent by blocking the timer ISR).
+    uint64_t timerISRCalls;          // Actual timer ISR entry count this session
 } StreamingStats;
 
 // Copies stats into *out inside a critical section (atomic snapshot)
